Limit the max number of switch ports based on hardware
Limit the max number of switch ports based on the hardware
specifications from TP-Link switches. The max switch ports is now 48,
matching the maximum in their hardware. When an actual device is
detected, it the device will be checked. These values are used for
static validation of the user's settings (e.g. default port).