Remove some default cards for turning on and off
Due to the name of the default onoff capability being named 'onoff',
there were some default cards added which did not make sense since
they described turning the entire device's power on and off and they
were used to turn the a switch port on and off. So these cards were
removed.

Renamed the switch from default to favorite.