Fix repair and increase polling frequency
Fixed the repair so that it retains the original settings in the UI.
Handle the case where the switch power is off when homey starts so
that the device reports as unavailable until it is discovered.
Increased the polling frequency without affecting the frequency of
re-login since login can cause a user on the device's web UI to be
logged off.