Add trigger card for when a switch port link changes
Added a trigger card for when a switch port link changes. The switch
port changes are polled on a best effort basis. The polling is
currently once per minute. Although there are some cases where it does
poll slighly less often based on the time it took to complete the
prior poll cycle.

This required some refactoring of the repair process. The repair
suspends the polling temporarily. The repair will wait for the polling
to stop and and will wait an additional 15 seconds in order to
decrease the liklihood that polling interferes with the repair
process.